Have updated patch link point to review, not commit message
Currently, when a patch is updated, the link directs to the commit
message, rather than the link to the relative cahnge. i.e.
X..Y/COMMIT_MSG rather than X..Y.

This change removes the COMMIT_MSG suffix, so that the link directs to
the relative change, rather to the commit message.